wip: create quality router
Currently, the function is erroring as it can't create a browser for some reason
