+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "name": "quality",
       "type": "debugpy",
       "request": "launch",
-      "module": "podium.quality",
+      "module": "podium.routers.quality",
       "cwd": "${workspaceFolder}/backend",
       "env": {
         "ENV_FOR_DYNACONF": "development"
